+++ src/layer/RLayer.tsx
@@ -31,8 +31,6 @@ export interface RLayerProps extends PropsWithChildren<unknown> {
     attributions?: string;
     /** A set of properties that can be accessed later by .get()/.getProperties() */
     properties?: Record<string, unknown>;
-    /** The layer will be reprojected if its projection is different than the map */
-    projection?: ProjectionLike;
     /** Called on every change */
     onChange?: (this: RLayer<RLayerProps>, e: BaseEvent) => void;
 }